// src/toolbar/button/buttons/button-action-enum.ts
export enum ButtonActionEnum {
  New = 'new',
  Add = 'add',
  Edit = 'edit',
  AppImport = 'app-import',
  Metadata = 'metadata',
  Delete = 'delete',
  Remove = 'remove',
  MoveUp = 'moveup',
  MoveDown = 'movedown',
  InstanceList = 'instance-list',
  Publish = 'publish',
  Replace = 'replace',
  ItemHistory = 'item-history',
  Layout = 'layout',
  TemplateDevelop = 'template-develop',
  TemplateQuery = 'template-query',
  TemplateSettings = 'template-settings',
  ContentItems = 'contentitems',
  ContentType = 'contenttype',
  App = 'app',
  AppSettings = 'app-settings',
  AppResources = 'app-resources',
  Zone = 'zone',
  Custom = 'custom',
  More = 'more',
}